Merged WisdomPage and WisdomEdit to share parent component and route
I moved all of WisdomPage's JSX into the newly created WisdomView component.
I turned WisdomPage into a parent componenet for WisdomView and WisdomEdit.
I removed the '/wisdom/edit' route.
WisdomView displays the content of the current wisdom.
WisdomEdit is the form for editing the current wisdom.
These components are toggled by state.
I moved WisdomEdit from the 'pages' foulder to the 'componenets' foulder because it is now a child componenet inside of a page.