fixed bug where setRenderHome was being triggered too early when registering a new user. onAuthStateChanged runs both when the page is refreshed and when a new user is being registered.
When registering a new user: the new if check ensures that renderHome is not switched to true until the displayName of the firebaseUser object has been updated. This is crucial because getDataToDisplay in the Home useEffect relies on displayName being set to get a successful return from fetchUserData.

When refreshing the page: the new if check allows renderHome to be set to true because the displayName will have already been set on the firebaseUser object.