Added a new context, UtilityContext, and accompanying custom hook, useUtility.
This context holds a global state, didDelete, which is used to trigger the getDataToDisplay function in the useEffect found in the Home.tsx component after the handleDelete function in WisdomView.tsx has been triggered.

This is my current duct tape way of solving the issue where the useEffect fetch code was not being run after a delete process has occurred. This solution is not perfect. I may have a better solution if I decide to store the 'storedWisdoms' state in a local context instead of in the useState inside of the Home.tsx component. I want to make that decision on a separate dedicated branch.