Fixed routing/rendering issue with WisdomPageWrapper & WisdomPage
Fixed issue where after navigating to WisdomPage then to UserAccount then logging out then logging back in WisdomPage would be visible on the '/' route instead of the Home component. My use of history.push() throughout route transitions was the culprit. Switching to use routerLink and routerDirection props solved the problem.

I also removed a bunch of console.log s and other code that I was using to troubleshoot the problem.