Setup project for testing Firebase security rules with jest and typescript.
I added the firebase SDK, as well as the firebase emulators, to my project to facilitate writing and testing security rules in VS code. I had to add a couple packages to get jest and typescript to work together. I have set up an initial securityRules.test.ts file that is functionaing appropriately.