Worked in cloud functions index.ts edited original function to match my firestore structure.
planned out an addition to the function for updating the 'createdBy'
field in wisdom docs with an updated username when a user changes their name.

considering splitting this funcntion into two functions:
one to enforce unique username on registration.
another to enforce unique username on username change as well as
handle updating the user's current docs with the new username.