Reordered onSubmit() in RegisterForm.tsx & Ran into bug
Successfully reordered my logic in onSubmit() in RegisterForm.tsx so that if the batch write fails the user is logged out and returned to the 'login_register' route.

I have run into a very frustrating bug where a check in my security rules is denying me permission to write, but only in production.
This check comes back as 'true' in the test emulator and in the security rules playground. To be continued...